+++ example/three/deepZoom.js
@@ -5,7 +5,7 @@ import {
 	OrthographicCamera,
 } from 'three';
 import { EnvironmentControls, TilesRenderer, CameraTransitionManager } from '3d-tiles-renderer';
-import { DeepZoomImagePlugin, UpdateOnChangePlugin } from '3d-tiles-renderer/plugins';
+import { UpdateOnChangePlugin, DeepZoomOverlay, GeneratedSurfacePlugin } from '3d-tiles-renderer/plugins';
 import { GUI } from 'three/addons/libs/lil-gui.module.min.js';
 
 let controls, scene, renderer;
@@ -14,8 +14,6 @@ let tiles, transition;
 const params = {
 
 	errorTarget: 1,
-	renderScale: 1,
-
 	orthographic: false,
 
 };
@@ -53,8 +51,12 @@ function init() {
 	} );
 
 	// tiles
-	tiles = new TilesRenderer( 'https://openseadragon.github.io/example-images/duomo/duomo.dzi' );
-	tiles.registerPlugin( new DeepZoomImagePlugin( { center: true } ) );
+	tiles = new TilesRenderer();
+	tiles.registerPlugin( new GeneratedSurfacePlugin( {
+		overlay: new DeepZoomOverlay( {
+			url: 'https://openseadragon.github.io/example-images/duomo/duomo.dzi',
+		} ),
+	} ) );
 	tiles.registerPlugin( new UpdateOnChangePlugin() );
 	tiles.fetchOptions.mode = 'cors';
 	tiles.lruCache.minSize = 900;
@@ -91,12 +93,11 @@ function init() {
 
 	} );
 
-	gui.add( params, 'errorTarget', 0, 100 ).onChange( () => {
+	gui.add( params, 'errorTarget', 1, 100 ).onChange( () => {
 
 		tiles.getPluginByName( 'UPDATE_ON_CHANGE_PLUGIN' ).needsUpdate = true;
 
 	} );
-	gui.add( params, 'renderScale', 0.1, 1.0, 0.05 ).onChange( v => renderer.setPixelRatio( v * window.devicePixelRatio ) );
 
 	gui.open();