+++ src/three/plugins/images/ImageFormatPlugin.js
@@ -25,6 +25,8 @@ export class ImageFormatPlugin {
 
 	constructor( options = {} ) {
 
+		console.warn( `${ this.constructor.name } has been deprecated. Use "GeneratedSurfacePlugin" & "ImageOverlayPlugin", instead.` );
+
 		const {
 			pixelSize = null,
 			center = false,