+++ src/three/plugins/index.js
@@ -14,8 +14,10 @@ export * from './LoadRegionPlugin.js';
 export * from './DebugTilesPlugin.js';
 
 // other formats
+export * from './images/GeneratedSurfacePlugin.js';
 export * from './images/DeepZoomImagePlugin.js';
 export * from './images/EPSGTilesPlugin.js';
+export * from './images/MVTOverlay.js';
 
 // gltf extensions
 export * from './gltf/GLTFCesiumRTCExtension.js';