+++ CHANGELOG.md
@@ -7,13 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - "GeneratedSurfacePlugin" for generating an ellipsoidal or planar surface based on an "ImageOverlay".
+- Improved support for ExternalTexture memory tracking.
+- StructuralMetadata: Throw an error in `getPropertyTableData` when non-matching argument types are passed.
+- MVTOverlay, PMTilesOverlay: Overlays for reading and rendering vector tilee formats as tile overlays.
 
 ### Fixed
 - ImageFormatPlugins: Fixed case where tile nodes could be created and never released.
 - Converted remaining "three/examples/jsm/" paths to "three/addons/"
 
 ### Changed
-- Deprecated "ImageFormatPlugin" and derivative plugins to reduce code redundancy. Use "GeneratedSurfacePlugin", instead.
+- Deprecated "ImageFormatPlugin" and derivative plugins to reduce code redundancy. Use "GeneratedSurfacePlugin" & "ImageOverlayPlugin", instead.
 
 ## [0.4.24] - 2026.04.10
 ### Added